Vandal Agent path calculations in order to help the A* agent calculate it's path better.

+++ HurricaneEvacuation/SimulatorEnvironment/ISettings.cs
@@ -8,6 +8,6 @@ namespace HurricaneEvacuation.SimulatorEnvironment
         IList<IAgent> Agents { get; }
         int Deadline { get; }
         double SlowDown { get; }
-        double WeightConstant { get; }
+        int WeightConstant { get; }
     }
 }
